src/main/java/com/example/demo/rest/book/BookController.java
package com.example.demo.rest.book;

import com.example.demo.model.dto.BookDto;
import com.example.demo.service.BookService;
import jakarta.validation.Valid;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/v1/books")
public class BookController {

    /*Didn't have enough time for error handling using ControllerAdvice*/

    @Autowired
    private BookService bookService;

    @GetMapping
    public List<BookDto> getAllBooks() {
        return bookService.findAll();
    }

    @GetMapping("/{id}")
    public ResponseEntity<BookDto> getBookById(@PathVariable Long id) {
        return bookService.findById(id)
                .map(ResponseEntity::ok)
                .orElse(ResponseEntity.notFound().build());
    }

    @PostMapping
    public BookDto createBook(@Valid @RequestBody BookDto book) {
        return bookService.save(book);
    }

    @PutMapping("/{id}")
    public ResponseEntity<BookDto> updateBook(@PathVariable Long id, @Valid @RequestBody BookDto bookDetails) {
        return bookService.findById(id)
                .map(book -> ResponseEntity.ok(bookService.save(bookDetails)))
                .orElse(ResponseEntity.notFound().build());
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Object> deleteBook(@PathVariable Long id) {
        return bookService.findById(id)
                .map(book -> {
                    bookService.deleteById(id);
                    return ResponseEntity.ok().build();
                })
                .orElse(ResponseEntity.notFound().build());
    }

    @PostMapping("/borrow/{id}")
    public ResponseEntity<BookDto> borrowBook(@PathVariable Long id) {
        try {
            return ResponseEntity.ok(bookService.borrowBook(id));
        } catch (Exception e) {
            return ResponseEntity.badRequest().body(null);
        }
    }

    @PostMapping("/return/{id}")
    public ResponseEntity<BookDto> returnBook(@PathVariable Long id) {
        try {
            return ResponseEntity.ok(bookService.returnBook(id));
        } catch (Exception e) {
            return ResponseEntity.badRequest().body(null);
        }
    }
}
